Preview quiz: show the correct points possible
Closes CNVS-3856

TEST PLAN
---- ----

  - create a quiz with 3 questions with points possible: 0.42, 0.23, 1
  - preview the quiz and get all the questions correctly
    - verify you see "1.65 out of 1.65"
  - retry with failing some questions
    - verify you get the correct score shown

Reviewed-on: https://gerrit.instructure.com/39295
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Caleb Guanzon <cguanzon@instructure.com>
Product-Review: Ahmad Amireh <ahmad@instructure.com>

+++ app/models/quizzes/quiz_submission.rb
@@ -240,7 +240,7 @@ class Quizzes::QuizSubmission < ActiveRecord::Base
   end
 
   def points_possible_at_submission_time
-    self.questions_as_object.map { |q| q[:points_possible].to_i }.compact.sum || 0
+    self.questions_as_object.map { |q| q[:points_possible].to_f }.compact.sum || 0
   end
 
   def questions
